Dynamic GUI and Refactored code
Made the GUI dynamically update the list of datasets within an N5 file. Refactored the code so that there is a stronger separation of GUI and functional code, this should both allow test cases to no longer be dependent on GUI libraries allowing Github actions to work, and make the code more clean.