GUI, Image Handling, and S3 Testing
Added an okay button to the GUI for users to use and authenticate their decision on a file to choose. Also added a remote files button that will be used to open remote files when that finally works. Changed the library that will be opening images in ImageJ from N5IJ to N5 because N5IJ causes some error not allowing Blosc compressed images to be opened. This will mean that some boiler plate code will have to be written, but at least it won't raise errors. Finally some testing with S3 capabilities have been played with recently. No full implementations have been made, but few attempts where given.