Fixed Refresh Row Selected -1 Bug
When refreshing the table, if a row was already selected then the selected row would have been -1 because the table would be set to empty before it gets filled again which seems to upset the table model. So do nothing if the row selected is some range outside of what's expected.